fix bug for clustered news

+++ src/main/java/com/ml/bus/service/MemoryService.java
@@ -26,9 +26,9 @@ public class MemoryService {
 	IClusterDAO clusterDAO;
 	
 	Map<String, String> categoryUrl;
-	int unCategorySize;
-	int lessKClusterSize;
-	int rightSize;
+	double unCategorySize;
+	double lessKClusterSize;
+	double rightSize;
 	
 	private Map<String, List<Cluster>> categoryClusters;
